README.md에 규칙 추가 작성
