+++ src/main/java/site/metacoding/humancloud/domain/resume/ResumeDao.java
@@ -2,9 +2,9 @@ package site.metacoding.humancloud.domain.resume;
 
 import java.util.List;
 
-import site.metacoding.humancloud.dto.request.resume.SaveDto;
-import site.metacoding.humancloud.dto.request.resume.UpdateDto;
-import site.metacoding.humancloud.dto.response.page.PagingDto;
+import site.metacoding.humancloud.dto.dummy.request.resume.SaveDto;
+import site.metacoding.humancloud.dto.dummy.request.resume.UpdateDto;
+import site.metacoding.humancloud.dto.dummy.response.page.PagingDto;
 
 public interface ResumeDao {
 	public void save(SaveDto saveDto);
