+++ src/main/java/site/metacoding/humancloud/dto/dummy/request/company/SaveDto.java
@@ -1,4 +1,4 @@
-package site.metacoding.humancloud.dto.request.company;
+package site.metacoding.humancloud.dto.dummy.request.company;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -6,7 +6,9 @@ import site.metacoding.humancloud.domain.company.Company;
 
 @Getter
 @Setter
-public class UpdateDto {
+public class SaveDto {
+	private Integer companyId;
+	private String companyUsername;
 	private String companyPassword;
 	private String companyName;
 	private String companyEmail;
@@ -15,7 +17,7 @@ public class UpdateDto {
 	private String companyLogo;
 
 	public Company toEntity(String companyLogo) {
-		return new Company(this.companyPassword, this.companyName, this.companyEmail, this.companyPhoneNumber,
-				this.companyAddress, companyLogo);
+		return new Company(this.companyUsername, this.companyPassword, this.companyName, this.companyEmail,
+				this.companyPhoneNumber, this.companyAddress, companyLogo);
 	}
 }
