+++ src/main/java/site/metacoding/humancloud/dto/dummy/request/company/UpdateDto.java
@@ -1,4 +1,4 @@
-package site.metacoding.humancloud.dto.request.company;
+package site.metacoding.humancloud.dto.dummy.request.company;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -6,9 +6,7 @@ import site.metacoding.humancloud.domain.company.Company;
 
 @Getter
 @Setter
-public class SaveDto {
-	private Integer companyId;
-	private String companyUsername;
+public class UpdateDto {
 	private String companyPassword;
 	private String companyName;
 	private String companyEmail;
@@ -17,7 +15,7 @@ public class SaveDto {
 	private String companyLogo;
 
 	public Company toEntity(String companyLogo) {
-		return new Company(this.companyUsername, this.companyPassword, this.companyName, this.companyEmail,
-				this.companyPhoneNumber, this.companyAddress, companyLogo);
+		return new Company(this.companyPassword, this.companyName, this.companyEmail, this.companyPhoneNumber,
+				this.companyAddress, companyLogo);
 	}
 }
