+++ src/main/java/site/metacoding/humancloud/service/ApplyService.java
@@ -9,7 +9,7 @@ import lombok.RequiredArgsConstructor;
 import site.metacoding.humancloud.domain.apply.ApplyDao;
 import site.metacoding.humancloud.domain.resume.Resume;
 import site.metacoding.humancloud.domain.resume.ResumeDao;
-import site.metacoding.humancloud.dto.request.apply.SaveDto;
+import site.metacoding.humancloud.dto.dummy.request.apply.SaveDto;
 
 @RequiredArgsConstructor
 @Service
