+++ src/main/java/site/metacoding/humancloud/service/RecruitService.java
@@ -16,9 +16,9 @@ import site.metacoding.humancloud.domain.company.Company;
 import site.metacoding.humancloud.domain.company.CompanyDao;
 import site.metacoding.humancloud.domain.recruit.Recruit;
 import site.metacoding.humancloud.domain.recruit.RecruitDao;
-import site.metacoding.humancloud.dto.request.recruit.SaveDto;
-import site.metacoding.humancloud.dto.response.page.PagingDto;
-import site.metacoding.humancloud.dto.response.recruit.CompanyRecruitDto;
+import site.metacoding.humancloud.dto.dummy.request.recruit.SaveDto;
+import site.metacoding.humancloud.dto.dummy.response.page.PagingDto;
+import site.metacoding.humancloud.dto.dummy.response.recruit.CompanyRecruitDto;
 
 @RequiredArgsConstructor
 @Service
